Implementation_Phases/Phase_D_Retriever/requirements.txt
langchain==0.1.16
langchain-community==0.0.33
sentence-transformers==2.7.0
chromadb==0.5.0
python-dotenv==1.0.1
